Test code refactor
